Set Variable Labels for Multiple Variables

Sets variable labels for multiple variables at once using named arguments. All variable names must exist in the survey data.

Usage

set_variable_labels(x, ...)

Arguments

x

A survey design object.

...

Named arguments where the name is the variable (unquoted) and the value is the label string. Supports !!! list splicing for programmatic use.

Value

The modified survey object, invisibly.

See also

Examples

df <- data.frame(age = 25:30, income = 1:6, wt = rep(1, 6))
d  <- survey_taylor(
  data = df,
  variables = list(
    ids = NULL, weights = "wt", strata = NULL,
    fpc = NULL, nest = FALSE, probs_provided = FALSE
  )
)
d <- set_variable_labels(
  d,
  age    = "Age in years",
  income = "Annual household income"
)

# Programmatic with list splicing
lbls <- list(age = "Age in years", income = "Annual household income")
d <- set_variable_labels(d, !!!lbls)